Выделение логики Crawler в операции
